requirements.txt
typer>=0.9.0
pydantic>=2.0.0
PyYAML>=6.0

# V2.0.0 GUI Dependencies
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pywebview>=4.4.0
requests>=2.31.0
